Move: 프로세서로 condition, service, client 이동
